NO-REF: Rename base mapping classes
