+++ mappings/csv.py
@@ -1,9 +1,9 @@
-from .core import Core
+from .base_mapping import BaseMapping
 from logger import createLog
 
 logger = createLog(__name__)
 
-class CSVMapping(Core):
+class CSVMapping(BaseMapping):
     def __init__(self, source, statics):
         super().__init__(source, statics)
