+++ mappings/sql.py
@@ -1,10 +1,10 @@
-from .core import Core
+from .base_mapping import BaseMapping
 from logger import createLog
 
 logger = createLog(__name__)
 
 
-class SQLMapping(Core):
+class SQLMapping(BaseMapping):
     def __init__(self, source, statics):
         super().__init__(source, statics)
