+++ processes/ingest/doab.py
@@ -7,7 +7,7 @@ import requests
 from ..core import CoreProcess
 from logger import createLog
 from mappings.doab import DOABMapping
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from managers import DOABLinkManager
 
 logger = createLog(__name__)
