+++ processes/ingest/loc.py
@@ -3,7 +3,7 @@ import os, requests
 from requests.exceptions import HTTPError, ConnectionError
 
 from ..core import CoreProcess
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from mappings.loc import LOCMapping
 from managers import WebpubManifest
 from logger import createLog
