+++ processes/ingest/u_of_m.py
@@ -5,7 +5,7 @@ from botocore.exceptions import ClientError
 
 from ..core import CoreProcess
 from urllib.error import HTTPError
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from mappings.UofM import UofMMapping
 from managers import WebpubManifest
 from logger import createLog
