+++ tests/unit/processes/file/test_fulfill_manifest_process.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from processes.file.fulfill_url_manifest import FulfillURLManifestProcess
 from tests.helper import TestHelpers
