+++ tests/unit/processes/ingest/test_chicago_isac_process.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from processes import ChicagoISACProcess
 from tests.helper import TestHelpers
