+++ tests/unit/processes/ingest/test_loc_process.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mappings.core import MappingError
+from mappings.base_mapping import MappingError
 from model import Record
 from processes import LOCProcess
 from tests.helper import TestHelpers
